refactor(sprint2): extraire auth_handlers + monitoring_handlers (étape 1/5)
Modules créés:
- src/bin/email_api_dir/auth_handlers.rs: auth_login, auth_register,
  auth_logout, auth_refresh, auth_oauth_*, api_2fa_verify,
  api_password_reset_*, api_patch_user_locale, make_session, TOTP helpers
- src/bin/email_api_dir/monitoring_handlers.rs: api_monitoring_*,
  api_security_*, parse_window, since_str, query types

main.rs: blocs correspondants remplacés par commentaires.
Déclarations dans main.rs: mod auth_handlers; mod monitoring_handlers;
                           pub use auth_handlers::*; pub use monitoring_handlers::*;

Note: main.rs peut avoir des conflits de noms à résoudre en CI.
