+++ src/bin/email_api_dir/main.rs
@@ -48,6 +48,12 @@ use serde::{Deserialize, Serialize};
 // Ne modifie AUCUN comportement tant que le flag est OFF (défaut).
 #[path = "admin_auth.rs"]
 mod admin_auth;
+mod auth_handlers;
+mod monitoring_handlers;
+
+pub use auth_handlers::*;
+pub use monitoring_handlers::*;
+
 use sha1::Sha1;
 
 use chrono::{DateTime, Utc};
@@ -102,60 +108,11 @@ struct MailEvent {
 
 type EventBus = broadcast::Sender<MailEvent>;
 
-// --- Auth types ---
+// --- Shared auth types (moved to auth_handlers.rs) ---
+// LoginRequest, RegisterRequest, OAuthCallbackQuery, TwoFactorVerifyRequest
+// PasswordResetRequestBody, PasswordResetConfirmBody are in auth_handlers module.
 
-#[derive(Deserialize)]
-struct LoginRequest {
-    email: String,
-    password: String,
-}
-
-#[derive(Deserialize)]
-struct RegisterRequest {
-    #[serde(default)]
-    first_name: String,
-    #[serde(default)]
-    last_name: String,
-    /// Alias optionnel → adresse alias@misfits.ai redirigeant vers prenom.nom@misfits.ai
-    #[serde(default)]
-    alias: Option<String>,
-    password: String,
-    #[serde(default)]
-    condition_accepted: bool,
-}
-
-#[derive(Deserialize)]
-struct OAuthCallbackQuery {
-    code: Option<String>,
-    state: Option<String>,
-}
-
-// --- 2FA types ---
-
-fn default_2fa_method() -> String {
-    "email".to_string()
-}
-
-#[derive(Deserialize)]
-struct TwoFactorVerifyRequest {
-    email: String,
-    code: String,
-    #[serde(default = "default_2fa_method")]
-    method: String,
-}
-
-// --- Password reset types ---
-
-#[derive(Deserialize)]
-struct PasswordResetRequestBody {
-    email: String,
-}
-
-#[derive(Deserialize)]
-struct PasswordResetConfirmBody {
-    token: String,
-    new_password: String,
-}
+// --- 2FA default helper (still needed by main.rs serde defaults) ---
 
 // --- Send queue types ---
 
@@ -234,28 +191,10 @@ struct AuthResponse {
     session: SessionResponse,
 }
 
-fn make_session(email: &str, display_name: &str) -> AuthResponse {
-    let now = Utc::now().timestamp_millis() as u64;
-    AuthResponse {
-        session: SessionResponse {
-            id: Uuid::new_v4().to_string(),
-            user: UserResponse {
-                id: Uuid::new_v4().to_string(),
-                email: email.to_string(),
-                display_name: display_name.to_string(),
-                role: "admin".to_string(),
-                two_factor_enabled: false,
-                created_at: Utc::now().to_rfc3339(),
-                updated_at: Utc::now().to_rfc3339(),
-            },
-            access_token: Uuid::new_v4().to_string(),
-            refresh_token: Uuid::new_v4().to_string(),
-            expires_at: now + 3600_000,
-            refresh_expires_at: now + 604_800_000,
-            issued_at: now,
-        },
-    }
-}
+// make_session, verify_totp, generate_totp_secret, generate_otp_code
+// → moved to auth_handlers module
+
+
 
 async fn persist_event(mongo: &mongodb::Client, event: &MailEvent) {
     let db = env::var("MONGODB_DATABASE").unwrap_or_else(|_| "mailserver".to_string());
@@ -344,79 +283,15 @@ use simple_smtp_server::monitoring::alerts::AlertConfig;
 use simple_smtp_server::monitoring::storage;
 use simple_smtp_server::security;
 
-fn parse_window(s: &str) -> chrono::Duration {
-    let s = s.trim();
-    if let Some(n) = s.strip_suffix('m').and_then(|n| n.parse::<i64>().ok()) {
-        chrono::Duration::minutes(n)
-    } else if let Some(n) = s.strip_suffix('h').and_then(|n| n.parse::<i64>().ok()) {
-        chrono::Duration::hours(n)
-    } else if let Some(n) = s.strip_suffix('d').and_then(|n| n.parse::<i64>().ok()) {
-        chrono::Duration::days(n)
-    } else {
-        chrono::Duration::minutes(15)
-    }
-}
+// parse_window, since_str, env_bool, monitoring query types, dns_txt_lookup
+// api_monitoring_*, SecurityAlertsQuery, api_security_*
+// → moved to monitoring_handlers module
 
-fn since_str(window: &str) -> String {
-    let dur = parse_window(window);
-    (Utc::now() - dur).to_rfc3339()
-}
 
-fn env_bool(name: &str, default: bool) -> bool {
-    match env::var(name) {
-        Ok(v) => matches!(
-            v.trim().to_ascii_lowercase().as_str(),
-            "1" | "true" | "yes" | "on"
-        ),
-        Err(_) => default,
-    }
-}
 
-#[derive(Deserialize)]
-struct MonitoringWindowQuery {
-    #[serde(default = "default_monitoring_window")]
-    window: String,
-}
-fn default_monitoring_window() -> String {
-    "15m".into()
-}
+// ↑ monitoring/security handlers moved to monitoring_handlers module ↑
 
-fn default_window() -> String {
-    "1h".into()
-}
 
-#[derive(Deserialize)]
-struct MonitoringEventsQuery {
-    status: Option<String>,
-    from: Option<String>,
-    to: Option<String>,
-    provider: Option<String>,
-    country: Option<String>,
-    since: Option<String>,
-    until: Option<String>,
-    message_id: Option<String>,
-    #[serde(default = "default_mon_page")]
-    page: u32,
-    #[serde(default = "default_mon_page_size")]
-    page_size: u32,
-}
-fn default_mon_page() -> u32 {
-    1
-}
-fn default_mon_page_size() -> u32 {
-    50
-}
-
-#[derive(Deserialize)]
-struct MonitoringLiveQuery {
-    message_id: Option<String>,
-}
-
-#[derive(Deserialize)]
-struct AdminWindowQuery {
-    #[serde(default = "default_window")]
-    window: String,
-}
 
 #[derive(Deserialize)]
 struct DeliverabilityDiagnosticsQuery {
@@ -465,495 +340,8 @@ async fn dns_txt_lookup(name: &str) -> Vec<String> {
 }
 
 /// GET /api/monitoring/summary?window=15m
-async fn api_monitoring_summary(
-    query: web::Query<MonitoringWindowQuery>,
-    mongo: web::Data<Arc<mongodb::Client>>,
-) -> impl Responder {
-    let since = since_str(&query.window);
-    let base_filter = doc! { "ts": { "$gte": &since } };
+// monitoring + security handlers → monitoring_handlers module
 
-    let total = storage::count_events(&mongo, base_filter.clone()).await;
-
-    let by_status_docs = storage::aggregate(&mongo, vec![
-        doc! { "$match": base_filter.clone() },
-        doc! { "$group": { "_id": "$status", "count": { "$sum": 1 }, "avg_ms": { "$avg": "$total_ms" } } },
-    ]).await;
-
-    let mut by_status = serde_json::Map::new();
-    let mut total_delivered = 0u64;
-    let mut total_bounced = 0u64;
-    let mut avg_total_ms_sum = 0f64;
-    let mut avg_count = 0u32;
-
-    for doc in &by_status_docs {
-        let status = doc.get_str("_id").unwrap_or("unknown").to_string();
-        let count = doc.get_i64("count").unwrap_or(0) as u64;
-        let avg_ms = doc.get_f64("avg_ms").unwrap_or(0.0);
-        if status == "delivered" {
-            total_delivered = count;
-        }
-        if status == "bounced" {
-            total_bounced = count;
-        }
-        avg_total_ms_sum += avg_ms * count as f64;
-        avg_count += count as u32;
-        by_status.insert(status, serde_json::json!(count));
-    }
-
-    let delivery_rate = if total > 0 {
-        total_delivered as f64 / total as f64
-    } else {
-        0.0
-    };
-    let bounce_rate = if total > 0 {
-        total_bounced as f64 / total as f64
-    } else {
-        0.0
-    };
-    let avg_total_ms = if avg_count > 0 {
-        avg_total_ms_sum / avg_count as f64
-    } else {
-        0.0
-    };
-    let p95 = storage::p95_total_ms(&mongo, base_filter.clone(), 1000).await;
-
-    // Average risk score
-    let risk_docs = storage::aggregate(
-        &mongo,
-        vec![
-            doc! { "$match": base_filter.clone() },
-            doc! { "$group": { "_id": null, "avg_risk": { "$avg": "$risk_score" } } },
-        ],
-    )
-    .await;
-    let avg_risk = risk_docs
-        .first()
-        .and_then(|d| d.get_f64("avg_risk").ok())
-        .unwrap_or(0.0);
-
-    HttpResponse::Ok().json(serde_json::json!({
-        "window": query.window,
-        "since": since,
-        "total_events": total,
-        "by_status": by_status,
-        "delivery_rate": (delivery_rate * 1000.0).round() / 1000.0,
-        "bounce_rate": (bounce_rate * 1000.0).round() / 1000.0,
-        "avg_total_ms": avg_total_ms.round(),
-        "p95_total_ms": p95,
-        "avg_risk_score": (avg_risk * 10.0).round() / 10.0,
-    }))
-}
-
-/// GET /api/monitoring/events?status=&from=&to=&provider=&country=&since=&until=&page=
-async fn api_monitoring_events(
-    query: web::Query<MonitoringEventsQuery>,
-    mongo: web::Data<Arc<mongodb::Client>>,
-) -> impl Responder {
-    let mut filter = doc! {};
-
-    if let Some(ref s) = query.status {
-        filter.insert("status", s);
-    }
-    if let Some(ref f) = query.from {
-        filter.insert("from", doc! { "$regex": f.as_str(), "$options": "i" });
-    }
-    if let Some(ref t) = query.to {
-        filter.insert("to", doc! { "$regex": t.as_str(), "$options": "i" });
-    }
-    if let Some(ref p) = query.provider {
-        filter.insert("company", doc! { "$regex": p.as_str(), "$options": "i" });
-    }
-    if let Some(ref c) = query.country {
-        filter.insert("country", c);
-    }
-    if let Some(ref m) = query.message_id {
-        filter.insert("message_id", m);
-    }
-
-    let mut ts_filter = doc! {};
-    if let Some(ref s) = query.since {
-        ts_filter.insert("$gte", s);
-    }
-    if let Some(ref u) = query.until {
-        ts_filter.insert("$lte", u);
-    }
-    if !ts_filter.is_empty() {
-        filter.insert("ts", ts_filter);
-    }
-
-    let total = storage::count_events(&mongo, filter.clone()).await;
-    let events = storage::query_events(&mongo, filter, query.page, query.page_size).await;
-    let has_more = (query.page * query.page_size) < total as u32;
-
-    HttpResponse::Ok().json(serde_json::json!({
-        "events": events,
-        "total": total,
-        "page": query.page,
-        "page_size": query.page_size,
-        "has_more": has_more,
-    }))
-}
-
-/// GET /api/monitoring/messages/{message_id}/trace
-async fn api_monitoring_trace(
-    path: web::Path<String>,
-    mongo: web::Data<Arc<mongodb::Client>>,
-) -> impl Responder {
-    let message_id = path.into_inner();
-    let filter = doc! { "message_id": &message_id };
-    let mut events = storage::query_events(&mongo, filter, 1, 200).await;
-    // Chronological order for trace view
-    events.sort_by(|a, b| a.ts.cmp(&b.ts));
-
-    if events.is_empty() {
-        return HttpResponse::NotFound().json(serde_json::json!({ "message": "Trace not found" }));
-    }
-
-    let status = events
-        .iter()
-        .rev()
-        .find(|e| {
-            matches!(
-                e.status,
-                monitoring::SmtpStatus::Delivered
-                    | monitoring::SmtpStatus::Bounced
-                    | monitoring::SmtpStatus::Failed
-                    | monitoring::SmtpStatus::Deferred
-            )
-        })
-        .or_else(|| events.last())
-        .map(|e| format!("{:?}", e.status))
-        .unwrap_or_default();
-    let total_ms = events.iter().filter_map(|e| e.total_ms).max();
-
-    HttpResponse::Ok().json(serde_json::json!({
-        "message_id": message_id,
-        "status": status,
-        "total_ms": total_ms,
-        "steps": events.len(),
-        "trace": events,
-    }))
-}
-
-/// GET /api/monitoring/bounces?window=24h
-async fn api_monitoring_bounces(
-    query: web::Query<MonitoringWindowQuery>,
-    mongo: web::Data<Arc<mongodb::Client>>,
-) -> impl Responder {
-    let since = since_str(&query.window);
-    let filter = doc! { "ts": { "$gte": &since }, "status": "bounced" };
-    let events = storage::query_events(&mongo, filter.clone(), 1, 100).await;
-    let total = storage::count_events(&mongo, filter).await;
-
-    let hard = events
-        .iter()
-        .filter(|e| matches!(e.bounce_type, Some(monitoring::BounceType::Hard)))
-        .count();
-    let soft = events
-        .iter()
-        .filter(|e| matches!(e.bounce_type, Some(monitoring::BounceType::Soft)))
-        .count();
-    let policy = events
-        .iter()
-        .filter(|e| matches!(e.bounce_type, Some(monitoring::BounceType::Policy)))
-        .count();
-
-    HttpResponse::Ok().json(serde_json::json!({
-        "window": query.window,
-        "since": since,
-        "total": total,
-        "hard": hard,
-        "soft": soft,
-        "policy": policy,
-        "bounces": events,
-    }))
-}
-
-/// GET /api/monitoring/providers/top?window=24h
-async fn api_monitoring_providers_top(
-    query: web::Query<MonitoringWindowQuery>,
-    mongo: web::Data<Arc<mongodb::Client>>,
-) -> impl Responder {
-    let since = since_str(&query.window);
-    let pipeline = vec![
-        doc! { "$match": { "ts": { "$gte": &since } } },
-        doc! { "$group": {
-            "_id": { "company": "$company", "datacenter": "$datacenter", "country": "$country" },
-            "count":     { "$sum": 1 },
-            "delivered": { "$sum": { "$cond": { "if": { "$eq": ["$status", "delivered"] }, "then": 1, "else": 0 } } },
-            "bounced":   { "$sum": { "$cond": { "if": { "$eq": ["$status", "bounced"] },   "then": 1, "else": 0 } } },
-            "avg_total_ms": { "$avg": "$total_ms" },
-            "avg_risk":     { "$avg": "$risk_score" },
-        }},
-        doc! { "$sort": { "count": -1 } },
-        doc! { "$limit": 20 },
-    ];
-    let docs = storage::aggregate(&mongo, pipeline).await;
-
-    let providers: Vec<serde_json::Value> = docs
-        .iter()
-        .map(|d| {
-            let id = d.get_document("_id").ok();
-            let company = id
-                .and_then(|i| i.get_str("company").ok())
-                .unwrap_or("unknown");
-            let datacenter = id
-                .and_then(|i| i.get_str("datacenter").ok())
-                .unwrap_or("unknown");
-            let country = id
-                .and_then(|i| i.get_str("country").ok())
-                .unwrap_or("unknown");
-            serde_json::json!({
-                "company":      company,
-                "datacenter":   datacenter,
-                "country":      country,
-                "count":        d.get_i64("count").unwrap_or(0),
-                "delivered":    d.get_i64("delivered").unwrap_or(0),
-                "bounced":      d.get_i64("bounced").unwrap_or(0),
-                "avg_total_ms": d.get_f64("avg_total_ms").unwrap_or(0.0).round(),
-                "avg_risk_score": d.get_f64("avg_risk").unwrap_or(0.0).round(),
-            })
-        })
-        .collect();
-
-    HttpResponse::Ok().json(serde_json::json!({
-        "window": query.window,
-        "since": since,
-        "providers": providers,
-    }))
-}
-
-/// GET /api/monitoring/live?message_id=<optional>  — SSE
-async fn api_monitoring_live(query: web::Query<MonitoringLiveQuery>) -> HttpResponse {
-    let filter_mid = query.message_id.clone();
-
-    let rx = match monitoring::get_bus() {
-        Some(tx) => tx.subscribe(),
-        None => {
-            return HttpResponse::ServiceUnavailable()
-                .body("Monitoring bus not initialized (SMTP_MONITORING_ENABLED=true required)");
-        }
-    };
-
-    let event_stream = stream::unfold(
-        (
-            rx,
-            filter_mid,
-            tokio::time::interval(std::time::Duration::from_secs(15)),
-        ),
-        |(mut rx, mid, mut hb)| async move {
-            loop {
-                tokio::select! {
-                    result = rx.recv() => {
-                        match result {
-                            Ok(event) => {
-                                if let Some(ref f) = mid {
-                                    if &event.message_id != f { continue; }
-                                }
-                                let data = serde_json::to_string(&event).unwrap_or_default();
-                                let chunk = format!("event: smtp_event\ndata: {}\nretry: 3000\n\n", data);
-                                return Some((
-                                    Ok::<web::Bytes, actix_web::Error>(web::Bytes::from(chunk)),
-                                    (rx, mid, hb),
-                                ));
-                            }
-                            Err(broadcast::error::RecvError::Lagged(n)) => {
-                                eprintln!("monitoring SSE: lagged {} events", n);
-                                continue;
-                            }
-                            Err(broadcast::error::RecvError::Closed) => return None,
-                        }
-                    }
-                    _ = hb.tick() => {
-                        return Some((
-                            Ok::<web::Bytes, actix_web::Error>(
-                                web::Bytes::from(": heartbeat\n\n")
-                            ),
-                            (rx, mid, hb),
-                        ));
-                    }
-                }
-            }
-        },
-    );
-
-    HttpResponse::Ok()
-        .content_type("text/event-stream")
-        .insert_header(("Cache-Control", "no-cache"))
-        .insert_header(("X-Accel-Buffering", "no"))
-        .insert_header(("Connection", "keep-alive"))
-        .streaming(event_stream)
-}
-
-/// GET /api/monitoring/alerts/active
-async fn api_monitoring_alerts_active(
-    query: web::Query<MonitoringWindowQuery>,
-    mongo: web::Data<Arc<mongodb::Client>>,
-) -> impl Responder {
-    let window_minutes = parse_window(&query.window).num_minutes();
-    let config = AlertConfig::default();
-    let alerts = monitoring::alerts::evaluate_alerts(&mongo, window_minutes, &config).await;
-
-    HttpResponse::Ok().json(serde_json::json!({
-        "window": query.window,
-        "alert_count": alerts.len(),
-        "alerts": alerts,
-    }))
-}
-
-// ---------------------------------------------------------------------------
-// Security endpoints
-// ---------------------------------------------------------------------------
-
-#[derive(serde::Deserialize)]
-struct SecurityAlertsQuery {
-    #[serde(default = "default_window")]
-    window: String,
-    severity: Option<String>,
-    tenant_id: Option<String>,
-}
-
-#[derive(serde::Deserialize)]
-struct SecurityIncidentsQuery {
-    #[serde(default = "one")]
-    page: u32,
-    #[serde(default = "twenty")]
-    page_size: u32,
-    tenant_id: Option<String>,
-    severity: Option<String>,
-}
-
-fn one() -> u32 {
-    1
-}
-fn twenty() -> u32 {
-    20
-}
-
-async fn api_security_alerts_active(
-    query: web::Query<SecurityAlertsQuery>,
-    mongo: web::Data<Arc<mongodb::Client>>,
-) -> impl Responder {
-    use simple_smtp_server::security::audit;
-
-    let limit = 100i64;
-    let mut alerts = audit::query_active_alerts(&mongo, limit).await;
-
-    // Optional filters
-    if let Some(ref sev) = query.severity {
-        let sev_lc = sev.to_lowercase();
-        alerts.retain(|a| format!("{:?}", a.severity).to_lowercase() == sev_lc);
-    }
-    if let Some(ref tid) = query.tenant_id {
-        alerts.retain(|a| a.tenant_id.as_deref() == Some(tid.as_str()));
-    }
-
-    HttpResponse::Ok().json(serde_json::json!({
-        "window": query.window,
-        "alert_count": alerts.len(),
-        "alerts": alerts,
-    }))
-}
-
-async fn api_security_incidents(
-    query: web::Query<SecurityIncidentsQuery>,
-    mongo: web::Data<Arc<mongodb::Client>>,
-) -> impl Responder {
-    use mongodb::bson::doc;
-    use simple_smtp_server::security::audit;
-
-    let mut filter = doc! {};
-    if let Some(ref tid) = query.tenant_id {
-        filter.insert("tenant_id", tid.as_str());
-    }
-    if let Some(ref sev) = query.severity {
-        filter.insert("severity", sev.as_str());
-    }
-
-    let alerts = audit::query_alerts(&mongo, filter, query.page, query.page_size).await;
-    HttpResponse::Ok().json(serde_json::json!({
-        "page": query.page,
-        "page_size": query.page_size,
-        "count": alerts.len(),
-        "alerts": alerts,
-    }))
-}
-
-async fn api_security_tenant_status(
-    path: web::Path<String>,
-    mongo: web::Data<Arc<mongodb::Client>>,
-) -> impl Responder {
-    use mongodb::bson::doc;
-    let tenant_id = path.into_inner();
-    let db = std::env::var("MONGODB_DATABASE").unwrap_or_else(|_| "mailserver".to_string());
-    let coll = mongo
-        .database(&db)
-        .collection::<mongodb::bson::Document>("tenant_state");
-
-    match coll.find_one(doc! { "tenant_id": &tenant_id }).await {
-        Ok(Some(doc)) => HttpResponse::Ok().json(serde_json::json!({
-            "tenant_id": tenant_id,
-            "state": doc,
-        })),
-        Ok(None) => HttpResponse::Ok().json(serde_json::json!({
-            "tenant_id": tenant_id,
-            "state": null,
-        })),
-        Err(e) => {
-            HttpResponse::InternalServerError().json(serde_json::json!({ "error": e.to_string() }))
-        }
-    }
-}
-
-async fn api_security_rollback(
-    path: web::Path<String>,
-    mongo: web::Data<Arc<mongodb::Client>>,
-) -> impl Responder {
-    use simple_smtp_server::security::remediation;
-    let alert_id = path.into_inner();
-    match remediation::rollback_remediation(&mongo, &alert_id).await {
-        Ok(()) => HttpResponse::Ok()
-            .json(serde_json::json!({ "rolled_back": true, "alert_id": alert_id })),
-        Err(e) => HttpResponse::BadRequest().json(serde_json::json!({ "error": e })),
-    }
-}
-
-async fn api_security_live(mongo: web::Data<Arc<mongodb::Client>>) -> impl Responder {
-    use futures_util::stream;
-    use simple_smtp_server::security;
-    use std::time::Duration;
-
-    let rx = match security::get_bus() {
-        Some(tx) => tx.subscribe(),
-        None => return HttpResponse::ServiceUnavailable().body("security bus unavailable"),
-    };
-
-    let stream = stream::unfold(
-        (
-            rx,
-            tokio::time::interval(Duration::from_secs(15)),
-            mongo.clone(),
-        ),
-        |(mut rx, mut hb, mongo)| async move {
-            tokio::select! {
-                Ok(alert) = rx.recv() => {
-                    let data = serde_json::to_string(&alert).unwrap_or_default();
-                    let msg = format!("event: security_alert\ndata: {}\nretry: 3000\n\n", data);
-                    Some((Ok::<_, actix_web::Error>(actix_web::web::Bytes::from(msg)), (rx, hb, mongo)))
-                }
-                _ = hb.tick() => {
-                    Some((Ok(actix_web::web::Bytes::from_static(b": heartbeat\n\n")), (rx, hb, mongo)))
-                }
-            }
-        },
-    );
-
-    HttpResponse::Ok()
-        .content_type("text/event-stream")
-        .insert_header(("Cache-Control", "no-cache"))
-        .insert_header(("X-Accel-Buffering", "no"))
-        .streaming(stream)
-}
 
 async fn api_admin_security_posture(
     query: web::Query<AdminWindowQuery>,
@@ -2059,925 +1447,8 @@ fn welcome_email_html(
     )
 }
 
-async fn auth_login(
-    req: web::Json<LoginRequest>,
-    req_http: actix_web::HttpRequest,
-    logic: web::Data<Arc<Logic>>,
-    mongo: web::Data<Arc<mongodb::Client>>,
-) -> impl Responder {
-    let locale = i18n::resolve_locale(&get_accept_language(&req_http), None);
-    // Try to authenticate against MongoDB
-    match logic.authenticate_user(&req.email, &req.password).await {
-        Ok(Some(user)) => {
-            let display = if user.mailbox.is_empty() {
-                req.email.clone()
-            } else {
-                user.mailbox.clone()
-            };
-            HttpResponse::Ok().json(make_session(&req.email, &display))
-        }
-        Ok(None) => {
-            let env_user = env::var("SMTP_USERNAME").unwrap_or_default();
-            let env_pass = env::var("SMTP_PASSWORD").unwrap_or_default();
-            if req.email == env_user || req.email == format!("{}@misfits.ai", env_user) {
-                if req.password == env_pass {
-                    return HttpResponse::Ok().json(make_session(&req.email, &env_user));
-                }
-            }
-            // Emit auth failure event for brute-force detection
-            let ip = req_ip_str(&req_http);
-            let ev = simple_smtp_server::security::AuthEvent::new(
-                simple_smtp_server::security::AuthEventKind::ApiLogin,
-                &ip,
-                false,
-            );
-            let mc = mongo.clone();
-            tokio::spawn(async move {
-                simple_smtp_server::security::log_auth_event(&mc, ev).await;
-            });
-            HttpResponse::Unauthorized().json(serde_json::json!({
-                "message": i18n::t(&locale, "error-login-invalid", &[])
-            }))
-        }
-        Err(e) => {
-            eprintln!("Auth error: {}", e);
-            let env_user = env::var("SMTP_USERNAME").unwrap_or_default();
-            let env_pass = env::var("SMTP_PASSWORD").unwrap_or_default();
-            if req.email == env_user || req.email == format!("{}@misfits.ai", env_user) {
-                if req.password == env_pass {
-                    return HttpResponse::Ok().json(make_session(&req.email, &env_user));
-                }
-            }
-            HttpResponse::Unauthorized().json(serde_json::json!({
-                "message": i18n::t(&locale, "error-login-invalid", &[])
-            }))
-        }
-    }
-}
+// auth_login..api_password_reset_confirm → auth_handlers module
 
-async fn auth_register(
-    req: web::Json<RegisterRequest>,
-    req_http: actix_web::HttpRequest,
-    logic: web::Data<Arc<Logic>>,
-    bus: web::Data<EventBus>,
-    mongo: web::Data<Arc<mongodb::Client>>,
-) -> impl Responder {
-    let locale = i18n::resolve_locale(&get_accept_language(&req_http), None);
-    if !req.condition_accepted {
-        return HttpResponse::BadRequest().json(serde_json::json!({
-            "code": "CONDITIONS_NOT_ACCEPTED",
-            "message": i18n::t(&locale, "error-conditions-required", &[])
-        }));
-    }
-    if req.password.len() < 8 {
-        return HttpResponse::BadRequest().json(serde_json::json!({
-            "code": "INVALID_PASSWORD",
-            "message": i18n::t(&locale, "error-password-too-short", &[])
-        }));
-    }
-
-    let local_part = match build_misfits_local(&req.first_name, &req.last_name) {
-        Some(l) => l,
-        None => {
-            return HttpResponse::BadRequest().json(serde_json::json!({
-                "code": "MISSING_IDENTITY",
-                "message": i18n::t(&locale, "error-name-required", &[])
-            }));
-        }
-    };
-
-    // Adresse principale : prenom.nom@misfits.ai
-    let primary_email = format!("{}@misfits.ai", local_part);
-    // Alias éventuel : alias@misfits.ai → redirige vers primary_email
-    let alias_email: Option<String> = req
-        .alias
-        .as_deref()
-        .map(str::trim)
-        .filter(|s| !s.is_empty())
-        .map(|a| normalize_segment(a))
-        .filter(|n| !n.is_empty() && *n != local_part)
-        .map(|n| format!("{}@misfits.ai", n));
-
-    let display_name = {
-        let first = req.first_name.trim();
-        let last = req.last_name.trim();
-        if first.is_empty() && last.is_empty() {
-            local_part.clone()
-        } else {
-            format!("{} {}", first, last).trim().to_string()
-        }
-    };
-
-    let password = req.password.clone();
-    let password_hash = match web::block(move || bcrypt::hash(&password, 12)).await {
-        Ok(Ok(h)) => h,
-        Ok(Err(e)) => {
-            eprintln!("bcrypt error: {}", e);
-            return HttpResponse::InternalServerError().json(serde_json::json!({
-                "code": "INTERNAL_ERROR",
-                "message": i18n::t(&locale, "error-account-creation-failed", &[])
-            }));
-        }
-        Err(e) => {
-            eprintln!("bcrypt task error: {}", e);
-            return HttpResponse::InternalServerError().json(serde_json::json!({
-                "code": "INTERNAL_ERROR",
-                "message": i18n::t(&locale, "error-account-creation-failed", &[])
-            }));
-        }
-    };
-
-    match logic
-        .create_user(&primary_email, &password_hash, "inbox")
-        .await
-    {
-        Ok(_) => {}
-        Err(e) => {
-            let msg = e.to_string();
-            if msg.contains("E11000") || msg.contains("duplicate key") {
-                return HttpResponse::Conflict().json(serde_json::json!({
-                    "code": "EMAIL_TAKEN",
-                    "message": i18n::t(&locale, "error-email-taken", &[("email", &primary_email)])
-                }));
-            }
-            eprintln!("Register error ({}): {}", primary_email, e);
-            return HttpResponse::InternalServerError().json(serde_json::json!({
-                "code": "INTERNAL_ERROR",
-                "message": i18n::t(&locale, "error-account-creation-failed", &[])
-            }));
-        }
-    }
-
-    // Créer l'alias → primary si fourni
-    if let Some(ref alias) = alias_email {
-        if let Err(e) = logic.create_alias(alias, &primary_email).await {
-            eprintln!(
-                "Alias creation error ({} → {}): {}",
-                alias, primary_email, e
-            );
-            // Non bloquant : le compte est créé, l'alias sera à recréer
-        }
-    }
-
-    // Email de bienvenue HTML avec support RTL
-    let welcome_subject = i18n::t(&locale, "email-welcome-subject", &[]);
-    let welcome_body = welcome_email_html(
-        &locale,
-        &display_name,
-        &primary_email,
-        alias_email.as_deref(),
-    );
-    let welcome = Email {
-        id: Uuid::new_v4().to_string(),
-        from: "noreply@misfits.ai".to_string(),
-        to: primary_email.clone(),
-        subject: welcome_subject,
-        body: welcome_body,
-        headers: vec![(
-            "Content-Type".to_string(),
-            "text/html; charset=utf-8".to_string(),
-        )],
-        flags: vec![],
-        sequence_number: 1,
-        uid: 1,
-        internal_date: mongodb::bson::DateTime::from_millis(Utc::now().timestamp_millis()),
-        dkim_signature: None,
-    };
-    if let Err(e) = logic.deliver_to_inbox(&local_part, &welcome).await {
-        eprintln!("Welcome email delivery error ({}): {}", primary_email, e);
-    } else {
-        emit_event(
-            &bus,
-            &mongo,
-            MailEvent {
-                id: Uuid::new_v4().to_string(),
-                kind: MailEventKind::Received,
-                user_id: local_part.clone(),
-                email_id: welcome.id.clone(),
-                subject: welcome.subject.clone(),
-                from: welcome.from.clone(),
-                to: welcome.to.clone(),
-                timestamp: Utc::now().to_rfc3339(),
-            },
-        )
-        .await;
-    }
-
-    let session = make_session(&primary_email, &display_name);
-    HttpResponse::Created()
-        .insert_header(("Content-Language", locale.as_str()))
-        .json(serde_json::json!({
-            "email": primary_email,
-            "alias": alias_email,
-            "session": session.session,
-            "locale": locale,
-        }))
-}
-
-#[derive(Deserialize)]
-struct PatchLocaleRequest {
-    locale: String,
-}
-
-async fn api_patch_user_locale(
-    req: actix_web::HttpRequest,
-    body: web::Json<PatchLocaleRequest>,
-    logic: web::Data<Arc<Logic>>,
-) -> impl Responder {
-    if !i18n::SUPPORTED_LOCALES.contains(&body.locale.as_str()) {
-        return HttpResponse::BadRequest().json(serde_json::json!({
-            "error": "unsupported_locale",
-            "supported": i18n::SUPPORTED_LOCALES,
-        }));
-    }
-    let username = resolve_user_id(&req);
-    match logic.update_user_locale(&username, &body.locale).await {
-        Ok(()) => HttpResponse::Ok().json(serde_json::json!({ "locale": body.locale })),
-        Err(e) => {
-            eprintln!("update_user_locale error: {}", e);
-            HttpResponse::InternalServerError().json(serde_json::json!({ "error": "internal" }))
-        }
-    }
-}
-
-async fn auth_logout() -> impl Responder {
-    HttpResponse::Ok().json(serde_json::json!({}))
-}
-
-async fn auth_refresh() -> impl Responder {
-    // Return a new mock session
-    HttpResponse::Ok().json(serde_json::json!({
-        "session": {
-            "id": Uuid::new_v4().to_string(),
-            "user": {
-                "id": Uuid::new_v4().to_string(),
-                "email": "admin@misfits.ai",
-                "displayName": "admin",
-                "role": "admin",
-                "twoFactorEnabled": false,
-                "createdAt": Utc::now().to_rfc3339(),
-                "updatedAt": Utc::now().to_rfc3339(),
-            },
-            "accessToken": Uuid::new_v4().to_string(),
-            "refreshToken": Uuid::new_v4().to_string(),
-            "expiresAt": (Utc::now().timestamp_millis() + 3600000) as u64,
-            "refreshExpiresAt": (Utc::now().timestamp_millis() + 604800000) as u64,
-            "issuedAt": Utc::now().timestamp_millis() as u64,
-        }
-    }))
-}
-
-async fn auth_oauth_start(path: web::Path<String>) -> impl Responder {
-    let provider_raw = path.into_inner();
-    let provider = match normalize_oauth_provider(&provider_raw) {
-        Some(p) => p,
-        None => {
-            return HttpResponse::BadRequest().json(serde_json::json!({
-                "message": "Unsupported OAuth provider."
-            }))
-        }
-    };
-
-    let state = Uuid::new_v4().to_string();
-    let callback_base = env::var("OAUTH_CALLBACK_BASE_URL")
-        .unwrap_or_else(|_| "https://mail.misfits.ai".to_string());
-
-    let auth_url = match provider.as_str() {
-        "github" => {
-            let client_id = match env::var("GITHUB_CLIENT_ID").ok().filter(|v| !v.is_empty()) {
-                Some(id) => id,
-                None => {
-                    eprintln!("OAuth start: GITHUB_CLIENT_ID is not set");
-                    return HttpResponse::InternalServerError().json(serde_json::json!({
-                        "message": "OAuth provider not configured."
-                    }));
-                }
-            };
-            let redirect_uri = format!(
-                "{}/api/auth/oauth/github/callback",
-                callback_base.trim_end_matches('/')
-            );
-            format!(
-                "https://github.com/login/oauth/authorize?client_id={}&redirect_uri={}&state={}&scope=user:email",
-                client_id,
-                urlencoding::encode(&redirect_uri),
-                state
-            )
-        }
-        _ => {
-            return HttpResponse::BadRequest().json(serde_json::json!({
-                "message": "Unsupported OAuth provider."
-            }))
-        }
-    };
-
-    HttpResponse::Found()
-        .insert_header(("Location", auth_url))
-        .finish()
-}
-
-async fn auth_oauth_callback(
-    path: web::Path<String>,
-    query: web::Query<OAuthCallbackQuery>,
-    logic: web::Data<Arc<Logic>>,
-) -> impl Responder {
-    let provider_raw = path.into_inner();
-    let provider = match normalize_oauth_provider(&provider_raw) {
-        Some(p) => p,
-        None => {
-            return HttpResponse::BadRequest().json(serde_json::json!({
-                "message": "Unsupported OAuth provider."
-            }))
-        }
-    };
-
-    let code = match query
-        .code
-        .as_ref()
-        .map(|v| v.trim())
-        .filter(|v| !v.is_empty())
-    {
-        Some(v) => v.to_string(),
-        None => {
-            return HttpResponse::BadRequest().json(serde_json::json!({
-                "message": "Missing OAuth authorization code."
-            }))
-        }
-    };
-
-    let _state = query.state.clone().unwrap_or_default();
-    let callback_base = env::var("OAUTH_CALLBACK_BASE_URL")
-        .unwrap_or_else(|_| "https://mail.misfits.ai".to_string());
-    let frontend_base =
-        env::var("FRONTEND_BASE_URL").unwrap_or_else(|_| "https://mail.misfits.ai".to_string());
-
-    let http_client = reqwest::Client::builder()
-        .user_agent("misfits-email-api/1.0")
-        .build()
-        .unwrap_or_else(|_| reqwest::Client::new());
-
-    let (subject, email, display) = match provider.as_str() {
-        "github" => {
-            let client_id = env::var("GITHUB_CLIENT_ID").unwrap_or_default();
-            let client_secret = env::var("GITHUB_CLIENT_SECRET").unwrap_or_default();
-            if client_id.is_empty() || client_secret.is_empty() {
-                eprintln!("OAuth callback: GITHUB_CLIENT_ID or GITHUB_CLIENT_SECRET is not set");
-                return HttpResponse::InternalServerError().json(serde_json::json!({
-                    "message": "OAuth provider not configured."
-                }));
-            }
-            let redirect_uri = format!(
-                "{}/api/auth/oauth/github/callback",
-                callback_base.trim_end_matches('/')
-            );
-
-            // Exchange code for access_token
-            let token_resp = http_client
-                .post("https://github.com/login/oauth/access_token")
-                .header("Accept", "application/json")
-                .json(&serde_json::json!({
-                    "client_id": client_id,
-                    "client_secret": client_secret,
-                    "code": code,
-                    "redirect_uri": redirect_uri,
-                }))
-                .send()
-                .await;
-
-            let access_token = match token_resp {
-                Ok(r) => match r.json::<GithubTokenResponse>().await {
-                    Ok(t) => {
-                        if let Some(err) = &t.error {
-                            eprintln!(
-                                "GitHub OAuth error: {} — {}",
-                                err,
-                                t.error_description.as_deref().unwrap_or("")
-                            );
-                            return HttpResponse::Unauthorized().json(serde_json::json!({
-                                "message": "OAuth authentication failed."
-                            }));
-                        }
-                        match t.access_token {
-                            Some(tok) if !tok.is_empty() => tok,
-                            _ => {
-                                eprintln!("GitHub OAuth: missing access_token in response");
-                                return HttpResponse::Unauthorized().json(serde_json::json!({
-                                    "message": "OAuth authentication failed."
-                                }));
-                            }
-                        }
-                    }
-                    Err(e) => {
-                        eprintln!("GitHub OAuth token parse error: {}", e);
-                        return HttpResponse::Unauthorized().json(serde_json::json!({
-                            "message": "OAuth authentication failed."
-                        }));
-                    }
-                },
-                Err(e) => {
-                    eprintln!("GitHub OAuth token request error: {}", e);
-                    return HttpResponse::Unauthorized().json(serde_json::json!({
-                        "message": "OAuth authentication failed."
-                    }));
-                }
-            };
-
-            // Fetch user profile
-            let user_resp = http_client
-                .get("https://api.github.com/user")
-                .bearer_auth(&access_token)
-                .header("Accept", "application/vnd.github+json")
-                .send()
-                .await;
-
-            let gh_user = match user_resp {
-                Ok(r) => match r.json::<GithubUser>().await {
-                    Ok(u) => u,
-                    Err(e) => {
-                        eprintln!("GitHub user profile parse error: {}", e);
-                        return HttpResponse::Unauthorized().json(serde_json::json!({
-                            "message": "OAuth authentication failed."
-                        }));
-                    }
-                },
-                Err(e) => {
-                    eprintln!("GitHub user profile request error: {}", e);
-                    return HttpResponse::Unauthorized().json(serde_json::json!({
-                        "message": "OAuth authentication failed."
-                    }));
-                }
-            };
-
-            let subject = gh_user.id.to_string();
-            let display = gh_user
-                .name
-                .as_deref()
-                .unwrap_or(&gh_user.login)
-                .to_string();
-
-            // Use profile email or fall back to emails endpoint
-            let email = if let Some(e) = gh_user.email.filter(|e| !e.is_empty()) {
-                e
-            } else {
-                let emails_resp = http_client
-                    .get("https://api.github.com/user/emails")
-                    .bearer_auth(&access_token)
-                    .header("Accept", "application/vnd.github+json")
-                    .send()
-                    .await;
-
-                match emails_resp {
-                    Ok(r) => match r.json::<Vec<GithubEmail>>().await {
-                        Ok(emails) => emails
-                            .into_iter()
-                            .find(|e| e.primary && e.verified)
-                            .map(|e| e.email)
-                            .unwrap_or_else(|| format!("github+{}@misfits.ai", subject)),
-                        Err(_) => format!("github+{}@misfits.ai", subject),
-                    },
-                    Err(_) => format!("github+{}@misfits.ai", subject),
-                }
-            };
-
-            (subject, email, display)
-        }
-        _ => {
-            return HttpResponse::BadRequest().json(serde_json::json!({
-                "message": "Unsupported OAuth provider."
-            }))
-        }
-    };
-
-    // Resolve the username: from MongoDB if available, else fall back to the OAuth email.
-    let username = match logic
-        .find_or_create_oauth_user(&provider, &subject, &email, Some(&display))
-        .await
-    {
-        Ok(user) => user.username,
-        Err(e) => {
-            // MongoDB unavailable/auth failure — degrade gracefully like auth_register does.
-            eprintln!("OAuth MongoDB error (degraded session): {}", e);
-            email.to_string()
-        }
-    };
-
-    let session = make_session(&username, &display);
-    let access_token = session.session.access_token.clone();
-    // Escape session JSON for safe inline embedding in a <script> block.
-    let session_json = serde_json::to_string(&session)
-        .unwrap_or_default()
-        .replace('\\', "\\\\")
-        .replace('`', "\\`");
-
-    // Bridge page: stores session in mfa.session (localStorage + sessionStorage)
-    // and sets the mfa_session cookie, then redirects to /, matching the FE auth contract.
-    let html = format!(
-        r#"<!DOCTYPE html><html><head><meta charset="utf-8">
-<title>Connexion…</title></head><body>
-<script>(function(){{
-  try {{
-    const s = JSON.parse(`{session_json}`);
-    localStorage.setItem('mfa.session', JSON.stringify(s));
-    sessionStorage.setItem('mfa.session', JSON.stringify(s));
-  }} catch(e) {{}}
-  window.location.replace('/mail');
-}})();</script>
-</body></html>"#,
-        session_json = session_json,
-    );
-
-    HttpResponse::Ok()
-        .content_type("text/html; charset=utf-8")
-        // mfa_session cookie matches what the backend sets on normal login
-        .insert_header((
-            "Set-Cookie",
-            format!(
-                "mfa_session={}; Path=/; HttpOnly; Secure; SameSite=Lax",
-                access_token
-            ),
-        ))
-        .body(html)
-}
-
-// --- Mail list (Phase A1, issue #166) -----------------------------------------
-
-#[derive(Deserialize)]
-struct EmailListQuery {
-    #[serde(default = "default_folder")]
-    folder: String,
-    #[serde(default = "default_page")]
-    page: u32,
-    #[serde(rename = "pageSize", default = "default_page_size")]
-    page_size: u32,
-}
-
-fn default_folder() -> String {
-    "inbox".to_string()
-}
-fn default_page() -> u32 {
-    1
-}
-fn default_page_size() -> u32 {
-    50
-}
-
-/// Canonical FE folder id → mailbox names to try in Mongo (SMTP historically used INBOX).
-fn folder_to_mailboxes(folder: &str) -> Vec<String> {
-    let f = folder.trim().to_ascii_lowercase();
-    match f.as_str() {
-        "inbox" => vec!["inbox".into(), "INBOX".into()],
-        "sent" => vec!["sent".into(), "SENT".into(), "Sent".into()],
-        "drafts" => vec!["drafts".into(), "DRAFTS".into(), "Drafts".into()],
-        "archive" => vec!["archive".into(), "ARCHIVE".into(), "Archive".into()],
-        "trash" => vec!["trash".into(), "TRASH".into(), "Trash".into()],
-        "spam" => vec!["spam".into(), "SPAM".into(), "Spam".into(), "Junk".into()],
-        other => vec![other.to_string(), other.to_ascii_uppercase()],
-    }
-}
-
-fn canonical_folder(folder: &str) -> Option<String> {
-    let f = folder.trim().to_ascii_lowercase();
-    match f.as_str() {
-        "inbox" | "sent" | "drafts" | "archive" | "trash" | "spam" => Some(f),
-        _ => None,
-    }
-}
-
-/// Resolve mailbox local-part. Convention: user_id = `admin` (not admin@misfits.ai).
-fn resolve_user_id(req: &actix_web::HttpRequest) -> String {
-    if let Some(id) = req
-        .headers()
-        .get("x-user-id")
-        .and_then(|v| v.to_str().ok())
-        .map(str::trim)
-        .filter(|s| !s.is_empty())
-    {
-        return id.to_string();
-    }
-    if let Some(email) = req
-        .headers()
-        .get("x-user-email")
-        .and_then(|v| v.to_str().ok())
-        .map(str::trim)
-        .filter(|s| !s.is_empty())
-    {
-        return email.split('@').next().unwrap_or(email).to_string();
-    }
-    env::var("SMTP_USERNAME").unwrap_or_else(|_| "admin".to_string())
-}
-
-// --- TOTP helpers (RFC 6238) ---
-
-fn compute_hotp(key: &[u8], counter: u64) -> u32 {
-    type HmacSha1 = Hmac<Sha1>;
-    let mut mac = HmacSha1::new_from_slice(key).expect("HMAC accepts any key size");
-    mac.update(&counter.to_be_bytes());
-    let result = mac.finalize().into_bytes();
-    let offset = (result[19] & 0x0f) as usize;
-    let code = ((result[offset] as u32 & 0x7f) << 24)
-        | ((result[offset + 1] as u32) << 16)
-        | ((result[offset + 2] as u32) << 8)
-        | (result[offset + 3] as u32);
-    code % 1_000_000
-}
-
-fn verify_totp(secret_b32: &str, code: &str) -> bool {
-    use constant_time_eq::constant_time_eq;
-    let s = secret_b32.to_uppercase();
-    let pad = s.len() % 8;
-    let padded = if pad == 0 {
-        s
-    } else {
-        format!("{}{}", s, "=".repeat(8 - pad))
-    };
-    let key = match BASE32.decode(padded.as_bytes()) {
-        Ok(k) => k,
-        Err(_) => return false,
-    };
-    let t = Utc::now().timestamp() / 30;
-    for delta in [-1i64, 0, 1] {
-        let counter = (t + delta).max(0) as u64;
-        let candidate = format!("{:06}", compute_hotp(&key, counter));
-        if constant_time_eq(candidate.as_bytes(), code.as_bytes()) {
-            return true;
-        }
-    }
-    false
-}
-
-fn generate_totp_secret() -> String {
-    BASE32.encode(Uuid::new_v4().as_bytes())
-}
-
-fn generate_otp_code() -> String {
-    let b = Uuid::new_v4();
-    let n = u32::from_be_bytes([
-        b.as_bytes()[0],
-        b.as_bytes()[1],
-        b.as_bytes()[2],
-        b.as_bytes()[3],
-    ]);
-    format!("{:06}", n % 1_000_000)
-}
-
-// --- 2FA verify (POST /api/auth/2fa/verify) ---
-
-async fn api_2fa_verify(
-    body: web::Json<TwoFactorVerifyRequest>,
-    mongo: web::Data<Arc<mongodb::Client>>,
-) -> impl Responder {
-    let db = mongo_db_name();
-
-    if body.method == "totp" {
-        let coll = mongo.database(&db).collection::<bson::Document>("users");
-        let local = body.email.split('@').next().unwrap_or(&body.email);
-        let user_doc = match coll
-            .find_one(doc! { "$or": [{ "username": local }, { "username": &body.email }] })
-            .await
-        {
-            Ok(Some(d)) => d,
-            Ok(None) => {
-                return HttpResponse::Unauthorized().json(serde_json::json!({
-                    "verified": false, "error": "User not found"
-                }))
-            }
-            Err(e) => {
-                return HttpResponse::InternalServerError()
-                    .json(serde_json::json!({ "error": e.to_string() }))
-            }
-        };
-
-        let totp_secret = match user_doc
-            .get_str("totp_secret")
-            .ok()
-            .filter(|s| !s.is_empty())
-        {
-            Some(s) => s.to_string(),
-            None => {
-                return HttpResponse::BadRequest().json(serde_json::json!({
-                    "verified": false, "error": "TOTP not configured for this user"
-                }))
-            }
-        };
-
-        if !verify_totp(&totp_secret, &body.code) {
-            return HttpResponse::Unauthorized().json(serde_json::json!({
-                "verified": false, "error": "Invalid TOTP code"
-            }));
-        }
-    } else {
-        // Email OTP: verify code stored in two_factor_codes collection
-        let coll = mongo
-            .database(&db)
-            .collection::<bson::Document>("two_factor_codes");
-        let now_ms = Utc::now().timestamp_millis();
-
-        match coll
-            .find_one(doc! {
-                "email": &body.email,
-                "code": &body.code,
-                "used": false,
-                "expires_at": { "$gt": bson::DateTime::from_millis(now_ms) },
-            })
-            .await
-        {
-            Ok(Some(d)) => {
-                if let Ok(oid) = d.get_object_id("_id") {
-                    let _ = coll
-                        .update_one(doc! { "_id": oid }, doc! { "$set": { "used": true } })
-                        .await;
-                }
-            }
-            Ok(None) => {
-                return HttpResponse::Unauthorized().json(serde_json::json!({
-                    "verified": false, "error": "Invalid or expired code"
-                }))
-            }
-            Err(e) => {
-                return HttpResponse::InternalServerError()
-                    .json(serde_json::json!({ "error": e.to_string() }))
-            }
-        }
-    }
-
-    let display = body
-        .email
-        .split('@')
-        .next()
-        .unwrap_or(&body.email)
-        .to_string();
-    let session = make_session(&body.email, &display);
-    HttpResponse::Ok().json(serde_json::json!({ "verified": true, "session": session.session }))
-}
-
-// --- Password reset (POST /api/auth/password-reset/request + /confirm) ---
-
-async fn api_password_reset_request(
-    body: web::Json<PasswordResetRequestBody>,
-    mongo: web::Data<Arc<mongodb::Client>>,
-    logic: web::Data<Arc<Logic>>,
-) -> impl Responder {
-    let db = mongo_db_name();
-    let email = body.email.trim().to_lowercase();
-    let local = email.split('@').next().unwrap_or(&email).to_string();
-
-    let users_coll = mongo.database(&db).collection::<bson::Document>("users");
-    let user_exists = matches!(
-        users_coll
-            .find_one(doc! { "$or": [{ "username": &local }, { "username": &email }] })
-            .await,
-        Ok(Some(_))
-    );
-
-    // Always return 200 to avoid user enumeration
-    if !user_exists {
-        return HttpResponse::Ok().json(serde_json::json!({
-            "message": "If the address is registered, a reset link has been sent."
-        }));
-    }
-
-    let token = Uuid::new_v4().to_string();
-    let expires_at = bson::DateTime::from_millis(Utc::now().timestamp_millis() + 3_600_000); // 1 hour
-
-    let tokens_coll = mongo
-        .database(&db)
-        .collection::<bson::Document>("password_reset_tokens");
-    // Invalidate any previous pending tokens for this email
-    let _ = tokens_coll
-        .update_many(
-            doc! { "email": &email, "used": false },
-            doc! { "$set": { "used": true } },
-        )
-        .await;
-
-    if let Err(e) = tokens_coll
-        .insert_one(doc! {
-            "token": &token,
-            "email": &email,
-            "expires_at": expires_at,
-            "used": false,
-        })
-        .await
-    {
-        eprintln!("password_reset_request insert error: {}", e);
-        return HttpResponse::InternalServerError()
-            .json(serde_json::json!({ "error": "Failed to create reset token" }));
-    }
-
-    let frontend_url =
-        env::var("FRONTEND_URL").unwrap_or_else(|_| "https://app.misfits.ai".to_string());
-    let reset_url = format!("{}/auth/reset-password?token={}", frontend_url, token);
-    let body_html = format!(
-        "<p>Click <a href=\"{url}\">here</a> to reset your password. This link expires in 1 hour.</p>\
-         <p>Or copy this link: {url}</p>",
-        url = reset_url
-    );
-    let reset_email = Email {
-        id: Uuid::new_v4().to_string(),
-        from: "noreply@misfits.ai".to_string(),
-        to: email.clone(),
-        subject: "Password Reset Request".to_string(),
-        body: body_html,
-        headers: vec![(
-            "Content-Type".to_string(),
-            "text/html; charset=utf-8".to_string(),
-        )],
-        flags: vec![],
-        sequence_number: 0,
-        uid: 0,
-        internal_date: bson::DateTime::from_millis(Utc::now().timestamp_millis()),
-        dkim_signature: None,
-    };
-    if let Err(e) = logic.deliver_to_inbox(&local, &reset_email).await {
-        eprintln!("password reset email delivery error: {}", e);
-    }
-
-    HttpResponse::Ok().json(serde_json::json!({
-        "message": "If the address is registered, a reset link has been sent."
-    }))
-}
-
-async fn api_password_reset_confirm(
-    body: web::Json<PasswordResetConfirmBody>,
-    mongo: web::Data<Arc<mongodb::Client>>,
-) -> impl Responder {
-    if body.new_password.len() < 8 {
-        return HttpResponse::BadRequest()
-            .json(serde_json::json!({ "error": "Password must be at least 8 characters" }));
-    }
-
-    let db = mongo_db_name();
-    let tokens_coll = mongo
-        .database(&db)
-        .collection::<bson::Document>("password_reset_tokens");
-    let now_ms = Utc::now().timestamp_millis();
-
-    let token_doc = match tokens_coll
-        .find_one(doc! {
-            "token": &body.token,
-            "used": false,
-            "expires_at": { "$gt": bson::DateTime::from_millis(now_ms) },
-        })
-        .await
-    {
-        Ok(Some(d)) => d,
-        Ok(None) => {
-            return HttpResponse::BadRequest()
-                .json(serde_json::json!({ "error": "Invalid or expired token" }))
-        }
-        Err(e) => {
-            return HttpResponse::InternalServerError()
-                .json(serde_json::json!({ "error": e.to_string() }))
-        }
-    };
-
-    let email = match token_doc.get_str("email") {
-        Ok(e) => e.to_string(),
-        Err(_) => {
-            return HttpResponse::InternalServerError()
-                .json(serde_json::json!({ "error": "Token data corrupted" }))
-        }
-    };
-    let local = email.split('@').next().unwrap_or(&email).to_string();
-
-    let new_password = body.new_password.clone();
-    let password_hash = match web::block(move || bcrypt::hash(&new_password, 12)).await {
-        Ok(Ok(h)) => h,
-        _ => {
-            return HttpResponse::InternalServerError()
-                .json(serde_json::json!({ "error": "Password hashing failed" }))
-        }
-    };
-
-    let users_coll = mongo.database(&db).collection::<bson::Document>("users");
-    match users_coll
-        .update_one(
-            doc! { "$or": [{ "username": &local }, { "username": &email }] },
-            doc! { "$set": { "password": &password_hash } },
-        )
-        .await
-    {
-        Ok(r) if r.matched_count == 0 => {
-            return HttpResponse::NotFound().json(serde_json::json!({ "error": "User not found" }))
-        }
-        Err(e) => {
-            return HttpResponse::InternalServerError()
-                .json(serde_json::json!({ "error": e.to_string() }))
-        }
-        _ => {}
-    }
-
-    // Invalidate the token
-    if let Ok(oid) = token_doc.get_object_id("_id") {
-        let _ = tokens_coll
-            .update_one(doc! { "_id": oid }, doc! { "$set": { "used": true } })
-            .await;
-    }
-
-    HttpResponse::Ok().json(serde_json::json!({ "message": "Password updated successfully" }))
-}
-
-#[derive(Serialize)]
 struct EmailAddressDto {
     name: String,
     address: String,
